requirements.txt
setuptools
wheel
ray[default]>=2.2.0
ray[air]>=2.2.0
ray[tune]>=2.2.0
pyarrow>=6.0.1
keras>=2.0.0
tensorflow>=2.0.0
numpy>=1.16
pandas>=1.3.0
scikit-learn>=1.1.2
scipy
insilicoseq
biopython==1.78
tqdm
cloudpickle